#1 Finalize basic MVP
Add volume
Change the range for the rate to match values from https://webplatformcourse.com/bonus/speech-synthesis-api/
Enabled voice lookup by taking the approach laid out in https://blog.monotonous.org/2021/11/15/speechSynthesis-getVoices/.
Fix voice saving.

It turns out we fire events in a loop but only twice. When we change an attribute the method call pushes the event to the server which saves it to the assigns then broadcasts the message to save it to the app.payload.
That save touches the model property which fortunately does not fire the change event again.
Taking either side out of the equation messes with the 2-way binding, i.e. the frontend would work but the changes wouldn't get saved as attributes for when we load the notebook, refresh the page, or reinitialize the runtime.